+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 05.07.2020
+### Added
+- TwinCAT 2 support!
+- TwinCAT 3 support for older versions than 4022!
+
+### Changed
+- Symbol info (`symbol`) is returned in addition to the data type (`type`) and value in `readSymbol()` and similar
+- Device info is read when connected
+
 ## [1.8.0] - 02.07.2020
 ### Added
 - New method `invokeRpcMethod()`